# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Credit Repair Pro API",
    "description": "Complete API for Credit Repair SaaS — client management, credit tracking, dispute generation, payments, compliance (CROA/FCRA/GLBA), and AI-powered analysis.",
    "version": "1.0.0",
    "contact": {
      "name": "TriExpert Services",
      "url": "https://triexpertservice.com"
    }
  },
  "servers": [
    { "url": "https://triexpertservice.com/api", "description": "Production" },
    { "url": "http://localhost:5000/api", "description": "Development" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Local JWT or Auth0 token"
      }
    },
...
  },
  "paths": {
...
    "/disputes/{id}": {
      "get": {
        "tags": ["Disputes"], "summary": "Get single dispute", "operationId": "getDispute",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Dispute details" }, "404": { "description": "Not found" } }
      }
    },
    "/ai-disputes/strategy/{creditItemId}": {
      "get": { "tags": ["AI Disputes"], "summary": "Get recommended dispute strategy", "operationId": "getStrategy", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "creditItemId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }, { "name": "bureau", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "Strategy recommendation" } } }
    },
    "/ai-disputes/strategies/overview": {
      "get": { "tags": ["AI Disputes"], "summary": "Get all available strategies", "operationId": "getStrategiesOverview", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "All strategies" } } }
    },
    "/ai-disputes/generate": {
      "post": { "tags": ["AI Disputes"], "summary": "Generate dispute letter via OpenAI", "operationId": "generateAIDispute", "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["creditItemId", "disputeType", "bureau"], "properties": { "creditItemId": { "type": "string", "format": "uuid" }, "disputeType": { "type": "string" }, "bureau": { "type": "string" }, "additionalDetails": { "type": "string" } } } } } }, "responses": { "200": { "description": "Generated letter" } } }
    },
    "/ai-disputes/save": {
      "post": { "tags": ["AI Disputes"], "summary": "Save generated letter as draft", "operationId": "saveAIDispute", "security": [{ "bearerAuth": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["creditItemId", "content", "disputeType", "bureau"], "properties": { "creditItemId": { "type": "string" }, "content": { "type": "string" }, "disputeType": { "type": "string" }, "bureau": { "type": "string" } } } } } }, "responses": { "201": { "description": "Draft saved" } } }
    },
    "/ai-disputes/drafts": {
      "get": { "tags": ["AI Disputes"], "summary": "Get user's dispute drafts & sent", "operationId": "getDrafts", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "Drafts and sent letters" } } }
    },
    "/ai-disputes/{id}": {
      "get": { "tags": ["AI Disputes"], "summary": "Get specific dispute letter", "operationId": "getAIDispute", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "Dispute letter" } } },
      "delete": { "tags": ["AI Disputes"], "summary": "Delete a dispute draft", "operationId": "deleteAIDispute", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "Draft deleted" } } }
    },
    "/ai-disputes/{id}/send": {
      "patch": { "tags": ["AI Disputes"], "summary": "Mark dispute as sent", "operationId": "markAIDisputeSent", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "Marked as sent" } } }
    },
...
  }
}